requirements.txt
# requirements.txt
flask==3.0.3
pymongo==4.10.1
redis==5.0.8
sentence-transformers==3.1.1
requests==2.32.3
scikit-learn==1.5.2
numpy==1.26.4
python-dotenv==1.0.1